LibWeb: Collapse the SVG paintable classes into kind dispatch
With the viewport geometry and committed path in the arena, the nine
SVG paintable classes and the SVGMaskable mixin carried nothing but
virtual dispatch that the arena's PaintableKind answers directly.
Delete them: the SVG type predicates become kind tests, the SVG
variant of compute_absolute_rect folds into the devirtualized base,
the mask and clip queries become kind-gated Paintable methods over the
former SVGMaskable logic in SVGMasking.cpp, and the host callbacks
reach the SVG DOM elements through dom_node() with kind gates instead
of downcasts. create_paintable ends at Paintable, PaintableWithLines,
InlinePaintable, and ViewportPaintable, and the caller-less
forms_unconnected_subtree and should_anti_alias hooks go with it.
