--- Libraries/LibWeb/Painting/SVGClipPaintable.h
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2024, MacDue <macdue@dueutil.tech>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#pragma once
-
-#include <LibWeb/Layout/Box.h>
-#include <LibWeb/Painting/SVGPaintable.h>
-
-namespace Web::Painting {
-
-class SVGClipPaintable : public SVGPaintable {
-public:
-    static NonnullRefPtr<SVGClipPaintable> create(Layout::Box const&);
-
-    bool forms_unconnected_subtree() const override
-    {
-        return true;
-    }
-
-protected:
-    SVGClipPaintable(Layout::Box const&);
-};
-
-}
