--- Libraries/LibWeb/Painting/SVGForeignObjectPaintable.h
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2024, Aliaksandr Kalenik <kalenik.aliaksandr@gmail.com>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#pragma once
-
-#include <LibWeb/Layout/BlockContainer.h>
-#include <LibWeb/Painting/PaintableWithLines.h>
-#include <LibWeb/Painting/SVGMaskable.h>
-
-namespace Web::Painting {
-
-class SVGForeignObjectPaintable final : public PaintableWithLines
-    , public SVGMaskable {
-public:
-    static NonnullRefPtr<SVGForeignObjectPaintable> create(Layout::BlockContainer const&);
-
-    virtual GC::Ptr<DOM::Node const> dom_node_of_svg() const override { return dom_node(); }
-    virtual Optional<CSSPixelRect> get_mask_area() const override { return get_svg_mask_area(); }
-    virtual Optional<Gfx::MaskKind> get_mask_type() const override { return get_svg_mask_type(); }
-    virtual Optional<CSSPixelRect> get_clip_area() const override { return get_svg_clip_area(); }
-
-protected:
-    SVGForeignObjectPaintable(Layout::BlockContainer const&);
-
-private:
-    virtual bool is_svg_foreign_object_paintable() const override { return true; }
-};
-
-template<>
-inline bool Paintable::fast_is<SVGForeignObjectPaintable>() const { return is_svg_foreign_object_paintable(); }
-
-}
