--- Libraries/LibWeb/Painting/SVGMaskPaintable.h
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2024, MacDue <macdue@dueutil.tech>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#pragma once
-
-#include <LibWeb/Layout/Box.h>
-#include <LibWeb/Painting/SVGGraphicsPaintable.h>
-
-namespace Web::Painting {
-
-class SVGMaskPaintable : public SVGGraphicsPaintable {
-public:
-    static NonnullRefPtr<SVGMaskPaintable> create(Layout::Box const&);
-
-    bool forms_unconnected_subtree() const override
-    {
-        // Masks should not be painted (i.e. reachable) unless referenced by another element.
-        return true;
-    }
-
-protected:
-    SVGMaskPaintable(Layout::Box const&);
-};
-
-}
