--- Libraries/LibWeb/Painting/SVGMaskable.cpp
@@ -1,128 +1,0 @@
-/*
- * Copyright (c) 2024-2026, Aliaksandr Kalenik <kalenik.aliaksandr@gmail.com>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#include <LibWeb/DOM/Document.h>
-#include <LibWeb/Layout/Box.h>
-#include <LibWeb/Page/Page.h>
-#include <LibWeb/Painting/Blending.h>
-#include <LibWeb/Painting/ResolvedCSSFilter.h>
-#include <LibWeb/Painting/SVGClipPaintable.h>
-#include <LibWeb/Painting/SVGForeignObjectPaintable.h>
-#include <LibWeb/Painting/SVGGraphicsPaintable.h>
-#include <LibWeb/Painting/SVGPaintable.h>
-#include <LibWeb/Painting/SVGPathPaintable.h>
-#include <LibWeb/SVG/SVGClipPathElement.h>
-#include <LibWeb/SVG/SVGMaskElement.h>
-#include <LibWeb/SVG/SVGSVGElement.h>
-
-namespace Web::Painting {
-
-static Layout::Box const* first_child_layout_box_of_kind(SVG::SVGGraphicsElement const& graphics_element, Layout::RustFFI::NodeKind kind)
-{
-    // NB: Called during painting.
-    if (!graphics_element.unsafe_layout_node())
-        return nullptr;
-    for (auto child = graphics_element.unsafe_layout_node()->first_child(); child; child = child->next_sibling()) {
-        if (child->kind() == kind)
-            return static_cast<Layout::Box const*>(child.ptr());
-    }
-    return nullptr;
-}
-
-static Layout::Box const* get_mask_box(SVG::SVGGraphicsElement const& graphics_element)
-{
-    return first_child_layout_box_of_kind(graphics_element, Layout::RustFFI::NodeKind::SVGMaskBox);
-}
-
-static Layout::Box const* get_clip_box(SVG::SVGGraphicsElement const& graphics_element)
-{
-    return first_child_layout_box_of_kind(graphics_element, Layout::RustFFI::NodeKind::SVGClipBox);
-}
-
-// The object bounding box covers the target's geometry alone. The paintable's border box is not
-// that: SVG layout inflates it by the visible stroke width. So, take the bounding box from the
-// target's geometry path, which carries no stroke.
-// AD-HOC: A group or a foreign object has no single geometry path, and we have no object bounding
-//         box for it — so its border box still stands in there.
-static CSSPixelRect target_user_space_object_bounding_box(Paintable const& target_paintable)
-{
-    if (auto const* committed_path = target_paintable.committed_svg_path())
-        return committed_path->bounding_box().to_type<CSSPixels>();
-    return target_paintable.absolute_border_box_rect();
-}
-
-Optional<CSSPixelRect> SVGMaskable::get_svg_mask_area() const
-{
-    auto const& graphics_element = as<SVG::SVGGraphicsElement const>(*dom_node_of_svg());
-    auto* mask_box = get_mask_box(graphics_element);
-    if (!mask_box)
-        return {};
-
-    auto target_paintable = as<Layout::Box>(*graphics_element.unsafe_layout_node()).paintable_box();
-    if (!target_paintable)
-        return {};
-
-    // Percentages in a userSpaceOnUse masking area resolve against the SVG viewport. The whole
-    // computation stays in the target's user space: that is the coordinate space of the mask node
-    // in the visual context tree.
-    Gfx::FloatSize viewport_size {};
-    if (auto const* viewport_paintable = nearest_svg_viewport_paintable_of(*mask_box))
-        viewport_size = svg_viewport_user_rect(*viewport_paintable).size();
-
-    auto target_object_bounding_box = target_user_space_object_bounding_box(*target_paintable);
-    return as<SVG::SVGMaskElement>(*mask_box->dom_node()).resolve_masking_area(target_object_bounding_box, viewport_size, Gfx::AffineTransform {});
-}
-
-Optional<CSSPixelRect> SVGMaskable::get_svg_clip_area() const
-{
-    auto const& graphics_element = as<SVG::SVGGraphicsElement const>(*dom_node_of_svg());
-    auto const* clip_box = get_clip_box(graphics_element);
-    if (!clip_box)
-        return {};
-
-    auto const& clip_paintable = as<SVGPaintable>(*clip_box->paintable_box());
-
-    // The area must cover the same space calculate_svg_clip_display_list paints the content in.
-    auto clip_path_transform = clip_paintable.layout_node().used_svg_element_transform();
-    if (as<SVG::SVGClipPathElement>(*clip_box->dom_node()).clip_path_units() == SVG::SVGUnits::ObjectBoundingBox)
-        clip_path_transform = object_bounding_box_content_units_transform().multiply(clip_path_transform);
-    // An empty clipping path will completely clip away the element that had the clip-path property applied.
-    return clip_paintable.clip_path_geometry_bounds(clip_path_transform).value_or(CSSPixelRect {});
-}
-
-static Gfx::MaskKind mask_type_to_gfx_mask_kind(CSS::MaskType mask_type)
-{
-    switch (mask_type) {
-    case CSS::MaskType::Alpha:
-        return Gfx::MaskKind::Alpha;
-    case CSS::MaskType::Luminance:
-        return Gfx::MaskKind::Luminance;
-    default:
-        VERIFY_NOT_REACHED();
-    }
-}
-
-Optional<Gfx::MaskKind> SVGMaskable::get_svg_mask_type() const
-{
-    auto const& graphics_element = as<SVG::SVGGraphicsElement const>(*dom_node_of_svg());
-    if (auto* mask_box = get_mask_box(graphics_element))
-        return mask_type_to_gfx_mask_kind(mask_box->mask_type());
-    return {};
-}
-
-Gfx::AffineTransform SVGMaskable::object_bounding_box_content_units_transform() const
-{
-    auto const& graphics_element = as<SVG::SVGGraphicsElement const>(*dom_node_of_svg());
-    auto target_paintable = as<Layout::Box>(*graphics_element.unsafe_layout_node()).paintable_box();
-    if (!target_paintable)
-        return {};
-    auto bounding_box = target_user_space_object_bounding_box(*target_paintable);
-    return Gfx::AffineTransform {}
-        .translate(bounding_box.location().to_type<float>())
-        .scale({ bounding_box.width().to_float(), bounding_box.height().to_float() });
-}
-
-}
