--- Libraries/LibWeb/Painting/SVGPaintable.cpp
@@ -1,80 +1,0 @@
-/*
- * Copyright (c) 2018-2022, Andreas Kling <andreas@ladybird.org>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#include <LibGfx/BoundingBox.h>
-#include <LibWeb/Painting/SVGPaintable.h>
-#include <LibWeb/SVG/SVGMaskElement.h>
-
-namespace Web::Painting {
-
-SVGPaintable::SVGPaintable(Layout::Box const& layout_box)
-    : Paintable(layout_box)
-{
-}
-
-Layout::Box const& SVGPaintable::layout_box() const
-{
-    return static_cast<Layout::Box const&>(layout_node());
-}
-
-// https://drafts.csswg.org/css-masking-1/#ClipPathElement
-bool SVGPaintable::contributes_to_clip_path() const
-{
-    // If a child element is made invisible by display or visibility it does not contribute to the clipping path.
-    return layout_node().visibility() == CSS::Visibility::Visible && !display().is_none();
-}
-
-// https://drafts.csswg.org/css-masking-1/#ClipPathElement
-Optional<CSSPixelRect> SVGPaintable::clip_path_geometry_bounds(Gfx::AffineTransform const& additional_transform) const
-{
-    if (!contributes_to_clip_path())
-        return {};
-
-    // When the clipPath element contains multiple child elements, the silhouettes of the child elements are
-    // logically OR'd together to create a single silhouette which is then used to restrict the region onto
-    // which paint can be applied. Thus, a point is inside the clipping path if it is inside any of the
-    // children of the clipPath.
-    Gfx::BoundingBox<CSSPixels> bounding_box;
-    for_each_child_of_type<Paintable>([&](Paintable const& child) {
-        auto const* svg_paintable = as_if<SVGPaintable>(child);
-        if (!svg_paintable)
-            return IterationDecision::Continue;
-
-        auto child_transform = Gfx::AffineTransform { additional_transform }.multiply(child.layout_node().used_svg_element_transform());
-        auto child_bounds = svg_paintable->clip_path_geometry_bounds(child_transform);
-        if (!child_bounds.has_value())
-            return IterationDecision::Continue;
-
-        bounding_box.add_point(child_bounds->left(), child_bounds->top());
-        bounding_box.add_point(child_bounds->right(), child_bounds->bottom());
-        return IterationDecision::Continue;
-    });
-
-    if (bounding_box.has_no_points())
-        return {};
-    return bounding_box.to_rect();
-}
-
-CSSPixelRect SVGPaintable::compute_absolute_rect() const
-{
-    // SVG content geometry lives in the user space of the nearest ancestor viewport, and layout
-    // places every box viewport-relative already, so no ancestor offsets accumulate.
-    for (auto const* ancestor = layout_box().parent(); ancestor; ancestor = ancestor->parent()) {
-        if (ancestor->is_svg_svg_box())
-            return { offset(), content_size() };
-    }
-    return Paintable::compute_absolute_rect();
-}
-
-Gfx::ShouldAntiAlias SVGPaintable::should_anti_alias() const
-{
-    auto shape_rendering = layout_node().shape_rendering();
-    if (first_is_one_of(shape_rendering, CSS::ShapeRendering::Optimizespeed, CSS::ShapeRendering::Crispedges))
-        return Gfx::ShouldAntiAlias::No;
-    return Gfx::ShouldAntiAlias::Yes;
-}
-
-}
