--- Libraries/LibWeb/Painting/SVGPatternPaintable.cpp
@@ -1,21 +1,0 @@
-/*
- * Copyright (c) 2026, Tim Ledbetter <tim.ledbetter@ladybird.org>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#include <LibWeb/Painting/SVGPatternPaintable.h>
-
-namespace Web::Painting {
-
-NonnullRefPtr<SVGPatternPaintable> SVGPatternPaintable::create(Layout::Box const& layout_box)
-{
-    return adopt_ref(*new SVGPatternPaintable(layout_box));
-}
-
-SVGPatternPaintable::SVGPatternPaintable(Layout::Box const& layout_box)
-    : SVGPaintable(layout_box)
-{
-}
-
-}
