--- Libraries/LibWeb/Painting/SVGPatternPaintable.h
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2026, Tim Ledbetter <tim.ledbetter@ladybird.org>
- *
- * SPDX-License-Identifier: BSD-2-Clause
- */
-
-#pragma once
-
-#include <LibWeb/Layout/Box.h>
-#include <LibWeb/Painting/SVGPaintable.h>
-
-namespace Web::Painting {
-
-class SVGPatternPaintable : public SVGPaintable {
-public:
-    static NonnullRefPtr<SVGPatternPaintable> create(Layout::Box const&);
-
-    bool forms_unconnected_subtree() const override
-    {
-        return true;
-    }
-
-protected:
-    SVGPatternPaintable(Layout::Box const&);
-};
-
-}
